requirements.txt
# Lista las dependencias necesarias para el proyecto.

tensorflow==2.18.0
numpy==2.0.2
Pillow==10.0.0